@@ -8,7 +8,7 @@ readme = "README.md"
 packages = [{include = "pydantic_async_validation"}]
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.9"
 pydantic = ">=2.0.0,<3.0.0"
 fastapi = {version = ">=0.100.0,<1.0.0", optional = true}
 pytest = ">=7.4,<9.0"
@@ -38,6 +38,9 @@ ignore = ["A001","A002","A003","ANN101","ANN102","ANN401","C901","N8","B008","F4
 "__init__.py" = ["F401"]
 "conftest.py" = ["S101","ANN","F401"]
 "test_*.py" = ["S101","ANN","F401"]
+
+[tool.pytest.ini_options]
+asyncio_default_fixture_loop_scope = "function"
 
 [build-system]
 requires = ["poetry-core"]
