@@ -1,5 +1,6 @@
+from collections.abc import Generator
 from contextlib import contextmanager
-from typing import Generator, List, Optional, Tuple, Union
+from typing import Optional, Union
 
 from fastapi.exceptions import RequestValidationError
 from pydantic import ValidationError
@@ -7,9 +8,9 @@ from pydantic_core import ErrorDetails
 
 
 def _prefix_request_errors(
-    prefix: Tuple[Union[int, str], ...],
-    errors: List[ErrorDetails],
-) -> List[ErrorDetails]:
+    prefix: tuple[Union[int, str], ...],
+    errors: list[ErrorDetails],
+) -> list[ErrorDetails]:
     """Add prefix to errors in preparation for request validation error conversion."""
 
     return [
@@ -25,7 +26,7 @@ def _prefix_request_errors(
 
 @contextmanager
 def ensure_request_validation_errors(
-    prefix: Optional[Union[Tuple[Union[int, str], ...], str]] = None,
+    prefix: Optional[Union[tuple[Union[int, str], ...], str]] = None,
 ) -> Generator[None, None, None]:
     """
     Converter for `ValidationError` to `RequestValidationError`.
