@@ -1,6 +1,6 @@
 from functools import wraps
 from inspect import Signature, signature
-from typing import Callable, List, Tuple, Union, cast
+from typing import Callable, Union, cast
 
 from pydantic import PydanticUserError
 from pydantic_core import ErrorDetails, InitErrorDetails, PydanticCustomError
@@ -163,9 +163,9 @@ def generic_model_validator_wrapper(
 
 
 def prefix_errors(
-    prefix: Tuple[Union[int, str], ...],
-    errors: Union[List[InitErrorDetails], List[ErrorDetails]],
-) -> List[InitErrorDetails]:
+    prefix: tuple[Union[int, str], ...],
+    errors: Union[list[InitErrorDetails], list[ErrorDetails]],
+) -> list[InitErrorDetails]:
     """
     Extend all errors passed as list to include an additional prefix.
 
